Place recommend login version add with release v1.1.5

+++ src/main/java/yeoksamstationexit1/recommend/repository/PriorityRepository.java
@@ -3,6 +3,6 @@ package yeoksamstationexit1.recommend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import yeoksamstationexit1.recommend.entity.Priority;
 
-public interface PriorityRepository extends JpaRepository<Priority, Integer> {
+public interface PriorityRepository extends JpaRepository<Priority, Long> {
 
 }
